Update to v8-3.5.8-9030 and fix remaining build issues.
Since v8 is built with scons, it doesn't seem to work well with the libtool
hooks in the ports build system.  Despite there being 2 shared libraries in
the port, I cannot get them to be listed correctly as SHARED_LIBS in the
Makefile or in the pkg/PFRAG.shared manifest.  However, I'm not sure that
even matters.