Correct shared library handling for lang/v8
Since v8 is built with scons, it doesn't use libtool.  So we cannot use the
usual libtool convenience code in the OpenBSD ports system to collect the
shared library information.  We note it manually in the SHARED_LIBS variable
and make sure the port version number is specified in the ${V} variable.
This is the key issue that was holding me up.  Nowhere in the documentation
does it mention you need to do this (probably because most things use
libtool) and trying to trace the mess of Perl and make that is the OpenBSD
ports system makes my eyes bleed.

Changed V8_SVNREV to R to follow consistency with V8_VERSION becoming V.
The build system does not seem to use R for anything already.

Also removed the CHECK_LIB_DEPENDS variable as that's a user-specified thing
that we do not want permanently in the port.

Added SHARED_ONLY=Yes since we do not want to be building static libraries
for v8.