Start work on devel/patchutils port.
This is a port of patchutils-0.3.2 from Tim Waugh <twaugh@redhat.com>
Patches are in place to let it work with the /usr/bin/patch and
/usr/bin/diff programs that ship with OpenBSD rather than requiring
GNU patch.

Still to do are patching the remaining failing test cases, which mostly
have to do with slight differences in the operation of patch(1) and
diff(1) on OpenBSD.  See regress.log for current details.