Finish top level Makefile rewrite, implement .ignore files.
The top level Makefile is now sporting shiny BSD syntax.  I have written
more targets so I can use make's built in control flow statements like ".if"
and ".for".  This results in more internal targets, but it's less embedded
shell that you have to clumsily chain together.

I also implemented a .ignore file.  If the 'publish' targets see a .ignore
file in a port, it won't be published.  This is useful to me so I can add
ports to this project that I am not ready to publish yet.